move router encodings into routers subdirectory
