+++ packages/sdk/src/swap/providers/velodrome/encoding/poolRouter.ts
@@ -1,19 +1,13 @@
 import type { Address, Hex, PublicClient } from 'viem'
-import { encodeFunctionData, erc20Abi } from 'viem'
 
 import type { SupportedChainId } from '@/constants/supportedChains.js'
-import type {
-  VelodromeChainConfig,
-  VelodromeRouterType,
-} from '@/swap/providers/velodrome/config.js'
+import type { VelodromeChainConfig } from '@/swap/providers/velodrome/config.js'
 import type { ResolvedPoolConfig } from '@/swap/providers/velodrome/types.js'
 import type { Asset } from '@/types/asset.js'
 import type { SwapPrice } from '@/types/swap/index.js'
-import type { TransactionData } from '@/types/transaction.js'
-import { buildApprovalTxIfNeeded } from '@/utils/approve.js'
 
-import { encodeCLSwap, getCLQuote } from './cl.js'
-import { encodeSwap, getQuote } from './v2.js'
+import { encodeCLSwap, getCLQuote } from './routers/cl.js'
+import { encodeSwap, getQuote } from './routers/v2.js'
 
 /** Internal result from pool-type-specific quoting */
 export interface PoolQuoteResult {
@@ -22,7 +16,7 @@ export interface PoolQuoteResult {
 }
 
 /**
- * Fetch a price quote for the given pool type.
+ * Fetch a price quote by routing to the correct pool type (v2 AMM or CL/Slipstream).
  * @returns Internal quote and provider context for the SwapQuote
  * @throws If CL pool requested on a chain without CL factory/quoter
  */
@@ -86,7 +80,7 @@ export async function fetchPoolQuote(
 }
 
 /**
- * Encode swap calldata for the given pool type.
+ * Encode swap calldata by routing to the correct pool type (v2 AMM or CL/Slipstream).
  * @returns Encoded calldata as hex string
  */
 export function encodePoolSwap(
@@ -126,44 +120,5 @@ export function encodePoolSwap(
     recipient: params.recipient,
     deadline: params.deadline,
     chainId: params.chainId,
-  })
-}
-
-/**
- * Build a token approval or transfer transaction for swap input.
- *
- * Universal Router uses a direct ERC20 transfer instead of approve+transferFrom.
- * This works because smart wallet batching (4337) bundles the transfer and swap
- * into a single atomic UserOperation — the router receives tokens before executing
- * the swap in the same transaction. The caller must already hold the tokens.
- *
- * Legacy routers (v2, leaf) use standard approve, approving only the deficit.
- */
-export async function buildTokenApproval(
-  token: Address,
-  router: Address,
-  routerType: VelodromeRouterType,
-  amount: bigint,
-  owner: Address,
-  publicClient: PublicClient,
-): Promise<TransactionData | undefined> {
-  if (routerType === 'universal') {
-    return {
-      to: token,
-      data: encodeFunctionData({
-        abi: erc20Abi,
-        functionName: 'transfer',
-        args: [router, amount],
-      }),
-      value: 0n,
-    }
-  }
-
-  return buildApprovalTxIfNeeded({
-    publicClient,
-    token,
-    owner,
-    spender: router,
-    amount,
   })
 }
