+++ packages/sdk/src/swap/providers/velodrome/encoding/routers/cl.ts
@@ -14,7 +14,7 @@ import {
   buildSwapPrice,
   resolveTokens,
   UNIVERSAL_ROUTER_MSG_SENDER,
-} from './helpers.js'
+} from '../helpers.js'
 
 // ─────────────────────────────────────────────────────────────────────────────
 // Quoting
